Refactor test framework and test cases
- Rename files for code fixes
- Reformat test cases with the file-scoped namespace declarations
- Add CodeChange.Of(string, Func<string, string>)
- Add CodeFixVerifier.NewCodeChange(string)
- Remove CodeFixVerifier.ReadCodeChange(string)
- Add CodeFixVerifier.VerifyDiagnosticAndFix(CodeChange, ...)
- Add CodeFixVerifier.VerifyFix(CodeChange, ...)
- Remove CodeFixVerifier.VerifyFix(string, string, bool)
