+++ BeliefCrucible/CodeChange.cs
@@ -11,4 +11,22 @@ namespace BeliefCrucible;
 /// </param>
 public record CodeChange(string Before, string After)
 {
+    /// <summary>
+    /// Creates a CodeChange instance based on the specified name and function.
+    /// </summary>
+    /// <param name="name">
+    /// The name of the code file.
+    /// </param>
+    /// <param name="readText">
+    /// A function to read the text of the code file.
+    /// </param>
+    /// <returns>
+    /// A new CodeChange instance.
+    /// </returns>
+    public static CodeChange Of(string name, Func<string, string> readText)
+    {
+        var code = readText(name);
+        var codeFix = readText($"{name}_Fixed");
+        return new(code, codeFix);
+    }
 }
