+++ CodeFixes.Test/.editorconfig
@@ -38,3 +38,6 @@ dotnet_diagnostic.SA1008.severity = none
 
 # CS1591: Missing XML comment for publicly visible type or member
 dotnet_diagnostic.CS1591.severity = none
+
+# SA1600: Elements should be documented
+dotnet_diagnostic.SA1600.severity = none
