+++ TestSuite/Naming/SingleTypeParameter/Code_Fixed.cs
