+++ TestSuite/Refactoring/StaticGenericClass/ReferencedCode.cs
@@ -1,13 +1,12 @@
 #pragma warning disable CS0693
 
-namespace StyleChecker.Test.Refactoring.StaticGenericClass
-{
-    public static class ReferencedCode<T>
-    {
-        public static void Method(T instance)
-        {
-        }
+namespace StyleChecker.Test.Refactoring.StaticGenericClass;
 
-        public static T Identity(T instance) => instance;
+public static class ReferencedCode<T>
+{
+    public static void Method(T instance)
+    {
     }
+
+    public static T Identity(T instance) => instance;
 }
