+++ TestSuite/Refactoring/StaticGenericClass/RenameReferencedCode_Fixed.cs
@@ -2,13 +2,13 @@
 
 namespace StyleChecker.Test.Refactoring.StaticGenericClass;
 
-public static class ReferencedCode<T>
+public static class ReferencedCode_2
 {
-    public static void Method(T instance)
+    public static void Method<T>(T instance)
     {
     }
 
-    public static T Identity(T instance) => instance;
+    public static T Identity<T>(T instance) => instance;
 }
 
 public sealed class ReferencedCode_1;
