+++ TestSuite/Refactoring/TypeClassParameter/Okay.cs
@@ -1,57 +1,54 @@
 #pragma warning disable CS8321
 
-namespace StyleChecker.Test.Refactoring.TypeClassParameter
+namespace StyleChecker.Test.Refactoring.TypeClassParameter;
+
+using System;
+
+public sealed class Okay
 {
-    using System;
-
-    public sealed class Okay
+    public void IncludesStaticClass()
     {
-        public void IncludesStaticClass()
+        void Print(Type t)
         {
-            void Print(Type t)
-            {
-                Console.WriteLine(t.FullName);
-            }
-
-            Print(typeof(string));
-            Print(typeof(StaticClass));
+            Console.WriteLine(t.FullName);
         }
 
-        public void NotAllArgumentsAreTypeofLocalFunction()
-        {
-            void Print(Type t)
-            {
-                Console.WriteLine(t.FullName);
-            }
-
-            Print(typeof(string));
-            Print(GetType());
-        }
-
-        public void NobodyInvokesLocalFunction()
-        {
-            void NeverUsed(Type t)
-                => Console.WriteLine(t.FullName);
-        }
-
-        public void NotAllArgumentsAreTypeof(Type type)
-        {
-            Console.WriteLine(type.FullName);
-        }
-
-        public void NobodyInvokes(Type type)
-        {
-            Console.WriteLine(type.FullName);
-        }
-
-        public void Invoke()
-        {
-            NotAllArgumentsAreTypeof(typeof(string));
-            NotAllArgumentsAreTypeof(GetType());
-        }
-
-        public static class StaticClass
-        {
-        }
+        Print(typeof(string));
+        Print(typeof(StaticClass));
     }
+
+    public void NotAllArgumentsAreTypeofLocalFunction()
+    {
+        void Print(Type t)
+        {
+            Console.WriteLine(t.FullName);
+        }
+
+        Print(typeof(string));
+        Print(GetType());
+    }
+
+    public void NobodyInvokesLocalFunction()
+    {
+        void NeverUsed(Type t)
+            => Console.WriteLine(t.FullName);
+    }
+
+    public void NotAllArgumentsAreTypeof(Type type)
+    {
+        Console.WriteLine(type.FullName);
+    }
+
+    public void NobodyInvokes(Type type)
+    {
+        Console.WriteLine(type.FullName);
+    }
+
+    public void Invoke()
+    {
+        NotAllArgumentsAreTypeof(typeof(string));
+        NotAllArgumentsAreTypeof(GetType());
+    }
+
+    public static class StaticClass;
 }
